==> Manuscript/scripts/03_data_visualization.qmd <==
---
title: "Data in the Wild--Visualization"
author: "Ellen Bledsoe"
date: today
format: gfm
editor: visual
---

This document produces data visualizations for the pre- and post-survey data set.

## Packages and Data

First, we load the packages that we will need.

```{r message=FALSE}
library(tidyverse)  # for data wrangling and plotting
library(ggstats)    # for plotting Likert data with ggplot2
```

We also need the cleaned dataset.

```{r message=FALSE}
data <- read_csv("../data_clean/data_cleaned.csv")

head(data)
```

## Pre- and Post-Survey Comparisons

We want to make plots that shown the distributions of the answers pre- and post-survey.

First, we have to wrangle the data so that there is a column with the question numbers (for faceting) and calculate proportions for each score.

```{r}
data_long <- data |> 
  # create one column with all of the questions and one column with scores
  pivot_longer(starts_with("Q"),
               names_to = "Question",
               values_to = "Score") |> 
  # convert survey type into a factor
  mutate(Survey_Type = factor(Survey_Type, levels = c("Post", "Pre"))) |> 
  # group by question to get proportions and means
  group_by(Question, Survey_Type, Score) |>
  # how many responses with this score
  summarise(Responses_Score = n()) |>
  # how many responses per question/survey type and calculate proportion
  mutate(Responses_Question = sum(Responses_Score),
         Proportion = Responses_Score / Responses_Question,
         Avg_Score = mean(Score),
         Score = factor(Score, levels = c(10:0))) |>
  ungroup()
```

Because Question 39 is on a different scale, we will need to plot it separately.

```{r}
data_long |> 
  filter(Question == "Q39") |> 
  ggplot(aes(x = Responses_Score,
             y = Survey_Type,
             fill = Score)) +
  # position = "likert" aligns the bars at the central value
  # this allows you to better see the shift in "weight" in answers
  geom_bar(stat = "identity", position = "likert") +
  geom_vline(aes(xintercept = 0), color = "lightgray") +
  geom_text(aes(label = round(Avg_Score, 1), group = Survey_Type), x = 1) +
  scale_fill_viridis_d(direction = 1) +
  theme_classic() 
```

Plot the rest of the questions.

```{r}
data_long |> 
  filter(Question != "Q39") |> 
  ggplot(aes(x = Responses_Score,
             y = Survey_Type,
             fill = Score)) +
  # position = "likert" aligns the bars at the central value
  # this allows you to better see the shift in "weight" in answers
  geom_bar(stat = "identity", position = "likert") +
  geom_vline(aes(xintercept = 0), color = "lightgray") +
  facet_wrap(~ Question) +
  geom_text(aes(label = round(Avg_Score, 1), group = Survey_Type), x = 1) +
  scale_fill_viridis_d(direction = 1) +
  theme_classic()
```

## Positive vs. Negative Shifts
